requirements.txt
fastapi
uvicorn
torch
torchvision
python-multipart
Jinja2
Pillow
imageio
numpy
